feat(utils folder): rename files to common naming style
