+++ src/features/apps/details/api.ts
@@ -19,7 +19,7 @@
  ********************************************************************************/
 
 import { getApiBase } from 'services/EnvironmentService'
-import { HttpClient } from 'utils/HttpClient'
+import { HttpClient } from 'utils/httpClient'
 import type { AppDetails } from './types'
 import i18next from 'i18next'
 import { getHeaders } from 'services/RequestService'
